+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Add player input latency measurement
 - Add player input latency display (only available in PROFILE builds)
+- Add packet send/receive artificial latency sliders (only available in PROFILE builds)
 
 ### Changed